Added AlertService && Implemented Pulley
